feat/refactor: user&refactor
users앱 추가 및 전체적인 settings refactor
