--- src/backend/boiler_analysis/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class BoilerAnalysisConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'boiler_analysis'
