+++ src/backend/boiler_analysis/urls.py
@@ -1,12 +1,34 @@
-from django.urls import path, include
-from rest_framework.routers import DefaultRouter
-from .views import AnalysisProjectViewSet, ModelConfigurationViewSet, AnalysisResultViewSet
+"""mysite URL Configuration
 
-router = DefaultRouter()
-router.register(r'projects', AnalysisProjectViewSet, basename='projects')
-router.register(r'models', ModelConfigurationViewSet, basename='models')
-router.register(r'results', AnalysisResultViewSet, basename='results')
+The `urlpatterns` list routes URLs to views. For more information please see:
+    https://docs.djangoproject.com/en/4.0/topics/http/urls/
+Examples:
+Function views
+    1. Add an import:  from my_app import views
+    2. Add a URL to urlpatterns:  path('', views.home, name='home')
+Class-based views
+    1. Add an import:  from other_app.views import Home
+    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
+Including another URLconf
+    1. Import the include() function: from django.urls import include, path
+    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
+"""
+from django.contrib import admin
+from django.urls import path, include
+from django.conf import settings
+from django.conf.urls.static import static
+from drf_spectacular.views import SpectacularAPIView, SpectacularSwaggerView, SpectacularRedocView
 
 urlpatterns = [
-    path('api/', include(router.urls)),
+    path('admin/', admin.site.urls),
+    path('api/', include('users.urls')),
+    path('api-auth/', include('rest_framework.urls')),
+    
+    # Swagger/OpenAPI 문서
+    path('api/schema/', SpectacularAPIView.as_view(), name='schema'),
+    path('api/docs/', SpectacularSwaggerView.as_view(url_name='schema'), name='swagger-ui'),
+    path('api/redoc/', SpectacularRedocView.as_view(url_name='schema'), name='redoc'),
 ]
+
+if settings.DEBUG:
+    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
