+++ src/backend/example/urls.py
@@ -1,0 +1,12 @@
+# from django.urls import path, include
+# from rest_framework.routers import DefaultRouter
+# from .views import AnalysisProjectViewSet, ModelConfigurationViewSet, AnalysisResultViewSet
+
+# router = DefaultRouter()
+# router.register(r'projects', AnalysisProjectViewSet, basename='projects')
+# router.register(r'models', ModelConfigurationViewSet, basename='models')
+# router.register(r'results', AnalysisResultViewSet, basename='results')
+
+# urlpatterns = [
+#     path('api/', include(router.urls)),
+# ]
